Moved the code into a singular .cpp file
The .h file method was not working for me so i just moved it into the main.cpp file. Decided on making this a different branch just because i wasnt sure if we wanted to do it the .h way or keep it all in one file. Let me know what we're thinking